First pass at installers

+++ server/src/test/java/com/antigravity/race/OverallStandingsTest.java
@@ -20,30 +20,35 @@ public class OverallStandingsTest {
         return rp;
     }
 
-    private Heat createHeat(int number, RaceParticipant p1, float laps1, float time1, RaceParticipant p2, float laps2, float time2) {
+    private Heat createHeat(int number, RaceParticipant p1, double laps1, double time1, RaceParticipant p2,
+            double laps2, double time2) {
         List<DriverHeatData> dhdList = new ArrayList<>();
-        
+
         DriverHeatData d1 = new DriverHeatData(p1);
         // Add laps to simulate average time
-        for(int i=0; i<laps1; i++) d1.addLap(time1/laps1); 
-        
+        for (int i = 0; i < laps1; i++)
+            d1.addLap(time1 / laps1);
+
         DriverHeatData d2 = new DriverHeatData(p2);
-        for(int i=0; i<laps2; i++) d2.addLap(time2/laps2);
+        for (int i = 0; i < laps2; i++)
+            d2.addLap(time2 / laps2);
 
         dhdList.add(d1);
         dhdList.add(d2);
-        
-        RaceScoring scoring = new RaceScoring(FinishMethod.Lap, 10, HeatRanking.LAP_COUNT, HeatRankingTiebreaker.FASTEST_LAP_TIME);
-        
+
+        RaceScoring scoring = new RaceScoring(FinishMethod.Lap, 10, HeatRanking.LAP_COUNT,
+                HeatRankingTiebreaker.FASTEST_LAP_TIME);
+
         return new Heat(number, dhdList, scoring);
     }
 
     @Test
     public void testSimpleRanking() {
         // Use FASTEST_LAP_TIME as tiebreaker
-        RaceScoring scoring = new RaceScoring(FinishMethod.Lap, 10, HeatRanking.LAP_COUNT, HeatRankingTiebreaker.FASTEST_LAP_TIME);
+        RaceScoring scoring = new RaceScoring(FinishMethod.Lap, 10, HeatRanking.LAP_COUNT,
+                HeatRankingTiebreaker.FASTEST_LAP_TIME);
         OverallStandings os = new OverallStandings(scoring);
-        
+
         RaceParticipant p1 = createDriver("D1", "id1");
         RaceParticipant p2 = createDriver("D2", "id2");
         List<RaceParticipant> drivers = new ArrayList<>();
@@ -52,7 +57,7 @@ public class OverallStandingsTest {
 
         List<Heat> heats = new ArrayList<>();
         // Heat 1: P1 wins (10 laps, 100s), P2 (9 laps, 100s)
-        heats.add(createHeat(1, p1, 10, 100.0f, p2, 9, 100.0f)); 
+        heats.add(createHeat(1, p1, 10, 100.0f, p2, 9, 100.0f));
 
         os.recalculate(drivers, heats);
 
@@ -64,23 +69,24 @@ public class OverallStandingsTest {
 
     @Test
     public void testDroppedHeats() {
-        RaceScoring scoring = new RaceScoring(FinishMethod.Lap, 10, HeatRanking.LAP_COUNT, HeatRankingTiebreaker.FASTEST_LAP_TIME);
+        RaceScoring scoring = new RaceScoring(FinishMethod.Lap, 10, HeatRanking.LAP_COUNT,
+                HeatRankingTiebreaker.FASTEST_LAP_TIME);
         OverallStandings os = new OverallStandings(scoring);
         os.setDroppedHeats(1);
-        
+
         RaceParticipant p1 = createDriver("D1", "id1");
         List<RaceParticipant> drivers = new ArrayList<>();
         drivers.add(p1);
 
         List<Heat> heats = new ArrayList<>();
-        // Note: For createHeat helper, P1 is first arg, P2 is second. 
+        // Note: For createHeat helper, P1 is first arg, P2 is second.
         // We reuse p1 for both slots to simplify helper usage or pass dummy p2.
         RaceParticipant dummy = createDriver("Dummy", "dummy");
-        
+
         // Heat 1: 10 laps (Good)
-        heats.add(createHeat(1, p1, 10, 100.0f, dummy, 0, 0)); 
+        heats.add(createHeat(1, p1, 10, 100.0f, dummy, 0, 0));
         // Heat 2: 5 laps (Bad heat - should be dropped)
-        heats.add(createHeat(2, p1, 5, 100.0f, dummy, 0, 0)); 
+        heats.add(createHeat(2, p1, 5, 100.0f, dummy, 0, 0));
         // Heat 3: 12 laps (Best heat)
         heats.add(createHeat(3, p1, 12, 100.0f, dummy, 0, 0));
